Rename gpu-query to gpu-control to better describe it's functions
