+++ daemon/gpu-control.h
